move test to lightning.pytorch
