+++ tests/tests_fabric/test_fabric.py
@@ -924,27 +924,3 @@ def test_grad_clipping(clip_val, max_norm):
             torch_model, torch_optimizer, max_norm=max_norm, norm_type=2.0, error_if_nonfinite=True
         )
 
-
-@pytest.mark.parametrize("algo", ["value", "norm"])
-def test_grad_clipping_lm(algo):
-    from lightning.pytorch.core.module import LightningModule
-    from lightning.pytorch.utilities.enums import GradClipAlgorithmType
-
-    class DummyLM(LightningModule):
-        def __init__(self):
-            super().__init__()
-            self.model = nn.Linear(1, 1)
-
-    fabric = Fabric()
-    orig_model = DummyLM()
-    model = fabric.setup(orig_model)
-
-    fabric.clip_gradients = Mock()
-
-    optimizer = Mock()
-    model.clip_gradients(optimizer, gradient_clip_val=1e-3, gradient_clip_algorithm=GradClipAlgorithmType(algo))
-
-    if algo == "value":
-        fabric.clip_gradients.assert_called_once_with(orig_model, optimizer, clip_val=1e-3, max_norm=None)
-    else:
-        fabric.clip_gradients.assert_called_once_with(orig_model, optimizer, clip_val=None, max_norm=1e-3)
